+++ test/gtest/unfold.hpp
@@ -35,15 +35,15 @@
 
 struct UnfoldTestCase
 {
-    size_t N;
-    size_t C;
-    size_t D;
-    size_t H;
-    size_t W;
-    std::vector<int64_t> kernelSize;
-    std::vector<int64_t> stride;
-    std::vector<int64_t> padding;
-    std::vector<int64_t> dilation;
+    uint64_t N;
+    uint64_t C;
+    uint64_t D;
+    uint64_t H;
+    uint64_t W;
+    std::vector<uint64_t> kernelSize;
+    std::vector<uint64_t> stride;
+    std::vector<uint64_t> padding;
+    std::vector<uint64_t> dilation;
     bool isContiguous = true;
     friend std::ostream& operator<<(std::ostream& os, const UnfoldTestCase& tc)
     {
@@ -64,40 +64,40 @@ struct UnfoldTestCase
         return os;
     }
 
-    std::vector<size_t> GetInput()
+    std::vector<uint64_t> GetInput()
     {
         if((N != 0) && (C != 0) && (D != 0) && (H != 0) && (W != 0))
         {
-            return std::vector<size_t>({N, C, D, H, W});
+            return std::vector<uint64_t>({N, C, D, H, W});
         }
         else if((N != 0) && (C != 0) && (H != 0) && (W != 0))
         {
-            return std::vector<size_t>({N, C, H, W});
+            return std::vector<uint64_t>({N, C, H, W});
         }
         else if((N != 0) && (C != 0) && (W != 0))
         {
-            return std::vector<size_t>({N, C, W});
+            return std::vector<uint64_t>({N, C, W});
         }
         else if((N != 0) && (W != 0))
         {
-            return std::vector<size_t>({N, W});
+            return std::vector<uint64_t>({N, W});
         }
         else if((N != 0))
         {
-            return std::vector<size_t>({N});
+            return std::vector<uint64_t>({N});
         }
         else
         {
             std::cout << "Error Input Tensor Lengths\n" << std::endl;
-            return std::vector<size_t>({0});
+            return std::vector<uint64_t>({0});
         }
     }
 
-    std::vector<size_t> ComputeStrides(std::vector<size_t> inputDim) const
+    std::vector<uint64_t> ComputeStrides(std::vector<uint64_t> inputDim) const
     {
         if(!isContiguous)
             std::swap(inputDim.front(), inputDim.back());
-        std::vector<size_t> strides(inputDim.size());
+        std::vector<uint64_t> strides(inputDim.size());
         strides.back() = 1;
         for(int i = inputDim.size() - 2; i >= 0; --i)
             strides[i] = strides[i + 1] * inputDim[i + 1];
@@ -136,30 +136,30 @@ protected:
         auto&& handle = get_handle();
         config        = GetParam();
 
-        std::vector<size_t> in_dims    = config.GetInput();
-        std::vector<size_t> in_strides = config.ComputeStrides(in_dims);
+        std::vector<uint64_t> in_dims    = config.GetInput();
+        std::vector<uint64_t> in_strides = config.ComputeStrides(in_dims);
 
         auto gen_value = [](auto...) { return prng::gen_descreet_uniform_sign<T>(1e-2, 100); };
         auto gen_zero  = [&](auto...) { return 0; };
         input          = tensor<T>{in_dims, in_strides}.generate(gen_value);
 
         int spatial_dim_size = in_dims.size() - 2;
-        const int64_t N      = static_cast<int64_t>(in_dims[0]);
-        const int64_t C      = static_cast<int64_t>(in_dims[1]);
-        int64_t P = 1, L = 1;
-        std::vector<int64_t> ls;
+        const uint64_t N     = static_cast<uint64_t>(in_dims[0]);
+        const uint64_t C     = static_cast<uint64_t>(in_dims[1]);
+        uint64_t P = 1, L = 1;
+        std::vector<uint64_t> ls;
         for(int i = 0; i < spatial_dim_size; ++i)
         {
             P *= config.kernelSize[i];
-            int64_t l = (in_dims[i + 2] + 2 * config.padding[i] -
-                         config.dilation[i] * (config.kernelSize[i] - 1) - 1) /
-                            config.stride[i] +
-                        1;
+            uint64_t l = (in_dims[i + 2] + 2 * config.padding[i] -
+                          config.dilation[i] * (config.kernelSize[i] - 1) - 1) /
+                             config.stride[i] +
+                         1;
             L *= l;
             ls.push_back(l);
         }
 
-        std::vector<size_t> out_dims{N, C * P, L};
+        std::vector<uint64_t> out_dims{N, C * P, L};
 
         output     = tensor<T>{out_dims}.generate(gen_zero);
         outputHost = tensor<T>{out_dims}.generate(gen_zero);
@@ -179,13 +179,13 @@ protected:
                                              output.desc,
                                              output_dev.get(),
                                              config.kernelSize.data(),
-                                             static_cast<int64_t>(config.kernelSize.size()),
+                                             static_cast<uint64_t>(config.kernelSize.size()),
                                              config.stride.data(),
-                                             static_cast<int64_t>(config.stride.size()),
+                                             static_cast<uint64_t>(config.stride.size()),
                                              config.padding.data(),
-                                             static_cast<int64_t>(config.padding.size()),
+                                             static_cast<uint64_t>(config.padding.size()),
                                              config.dilation.data(),
-                                             static_cast<int64_t>(config.dilation.size()));
+                                             static_cast<uint64_t>(config.dilation.size()));
 
         cpu_unfold_fwd_4d<T>(
             input, outputHost, config.kernelSize, config.stride, config.padding, config.dilation);
@@ -224,8 +224,8 @@ protected:
         auto&& handle = get_handle();
         config        = GetParam();
 
-        std::vector<size_t> in_dims    = config.GetInput();
-        std::vector<size_t> in_strides = config.ComputeStrides(in_dims);
+        std::vector<uint64_t> in_dims    = config.GetInput();
+        std::vector<uint64_t> in_strides = config.ComputeStrides(in_dims);
 
         auto gen_value = [](auto...) { return prng::gen_descreet_uniform_sign<T>(1e-2, 100); };
         auto gen_zero  = [&](auto...) { return 0; };
@@ -233,22 +233,22 @@ protected:
         dinputHost     = tensor<T>{in_dims, in_strides}.generate(gen_zero);
 
         int spatial_dim_size = in_dims.size() - 2;
-        const int64_t N      = static_cast<int64_t>(in_dims[0]);
-        const int64_t C      = static_cast<int64_t>(in_dims[1]);
-        int64_t P = 1, L = 1;
-        std::vector<int64_t> ls;
+        const uint64_t N     = static_cast<uint64_t>(in_dims[0]);
+        const uint64_t C     = static_cast<uint64_t>(in_dims[1]);
+        uint64_t P = 1, L = 1;
+        std::vector<uint64_t> ls;
         for(int i = 0; i < spatial_dim_size; ++i)
         {
             P *= config.kernelSize[i];
-            int64_t l = (static_cast<int64_t>(in_dims[i + 2]) + 2 * config.padding[i] -
-                         config.dilation[i] * (config.kernelSize[i] - 1) - 1) /
-                            config.stride[i] +
-                        1;
+            uint64_t l = (static_cast<uint64_t>(in_dims[i + 2]) + 2 * config.padding[i] -
+                          config.dilation[i] * (config.kernelSize[i] - 1) - 1) /
+                             config.stride[i] +
+                         1;
             L *= l;
             ls.push_back(l);
         }
 
-        std::vector<size_t> out_dims{N, C * P, L};
+        std::vector<uint64_t> out_dims{N, C * P, L};
 
         doutput = tensor<T>{out_dims}.generate(gen_value);
 
@@ -267,13 +267,13 @@ protected:
                                               doutput.desc,
                                               doutput_dev.get(),
                                               config.kernelSize.data(),
-                                              static_cast<int64_t>(config.kernelSize.size()),
+                                              static_cast<uint64_t>(config.kernelSize.size()),
                                               config.stride.data(),
-                                              static_cast<int64_t>(config.stride.size()),
+                                              static_cast<uint64_t>(config.stride.size()),
                                               config.padding.data(),
-                                              static_cast<int64_t>(config.padding.size()),
+                                              static_cast<uint64_t>(config.padding.size()),
                                               config.dilation.data(),
-                                              static_cast<int64_t>(config.dilation.size()));
+                                              static_cast<uint64_t>(config.dilation.size()));
 
         cpu_unfold_bwd_4d<T>(
             dinputHost, doutput, config.kernelSize, config.stride, config.padding, config.dilation);
